HCK-9614: fix an issue with config loading
