+++ forward_engineering/isDropInStatements.js
@@ -1,0 +1,35 @@
+const { initPluginConfiguration } = require('../helpers/levelConfigHelper');
+const { sortUdt, prepareDefinitions } = require('./helpers/udtHelper');
+
+function isDropInStatements(data, logger, callback, app) {
+	try {
+		initPluginConfiguration(data.pluginConfiguration, logger);
+
+		let result;
+		const { udtTypeMap, modelDefinitions, externalDefinitions } = prepareDefinitions(data);
+
+		if (data.level === 'container') {
+			data = { ...data, udtTypeMap, modelDefinitions, externalDefinitions };
+			result = data.entities
+				.map(entityId => {
+					const jsonSchema = JSON.parse(data.jsonSchema[entityId]);
+					data.internalDefinitions = sortUdt(JSON.parse(data.internalDefinitions[entityId]));
+					return isDropInStatements(jsonSchema, data.udtTypeMap, data);
+				})
+				.some(Boolean);
+		} else if (data.level === 'entity') {
+			const jsonSchema = JSON.parse(data.jsonSchema);
+			const internalDefinitions = sortUdt(JSON.parse(data.internalDefinitions));
+			data = { ...data, udtTypeMap, modelDefinitions, externalDefinitions, jsonSchema, internalDefinitions };
+			result = isDropInStatements(data.jsonSchema, data.udtTypeMap, data);
+		}
+
+		callback(null, result);
+	} catch (e) {
+		callback({ message: e.message, stack: e.stack });
+	}
+}
+
+module.exports = {
+	isDropInStatements,
+};
