+++ helpers/levelConfigHelper.js
@@ -1,50 +1,33 @@
-const fs = require('fs');
-const path = require('path');
+let pluginConfiguration = {};
+let logger = {};
+const descriptors = {};
 
-const getConfig = pathToConfig => {
+const initPluginConfiguration = (config, appLogger) => {
+	logger = appLogger;
+	pluginConfiguration = config || {};
+};
+
+const getTypeConfig = typeName => {
+	if (descriptors[typeName]) {
+		return descriptors[typeName];
+	}
+
 	try {
-		const filePath = path.join(__dirname, '..', pathToConfig);
-		const fileContent = fs.readFileSync(filePath);
+		descriptors[typeName] = require(`../types/${typeName}.json`);
 
-		return JSON.parse(fileContent.toString().replace(/\/\*[\s\S]*?\*\//g, ''));
+		return descriptors[typeName];
 	} catch (e) {
 		return {};
 	}
 };
 
-const cacheResult = method => {
-	let result;
+const getFieldLevelConfig = type => pluginConfiguration.fieldLevelConfig;
 
-	return (...args) => {
-		if (!result) {
-			result = method(...args);
-		}
-
-		return result;
-	};
-};
-
-const getFieldLevelConfig = cacheResult(() =>
-	getConfig(path.join('properties_pane', 'field_level', 'fieldLevelConfig.json')),
-);
-const getEntityLevelConfig = cacheResult(() =>
-	getConfig(path.join('properties_pane', 'entity_level', 'entityLevelConfig.json')),
-);
-
-const getTypesConfig = cacheResult(() => {
-	const getName = typeFile => typeFile.replace(/\.json/, '');
-	const typeDir = path.join(__dirname, '..', 'types');
-	const types = fs.readdirSync(typeDir);
-
-	return types.reduce((typesMap, fileName) => {
-		typesMap[getName(fileName)] = getConfig(path.join('types', fileName));
-
-		return typesMap;
-	}, {});
-});
+const getEntityLevelConfig = () => pluginConfiguration.entityLevelConfig;
 
 module.exports = {
-	getEntityLevelConfig,
+	initPluginConfiguration,
+	getTypeConfig,
 	getFieldLevelConfig,
-	getTypesConfig,
+	getEntityLevelConfig,
 };
